requirements.txt
flask
gunicorn
psutil
psycopg2-binary

